chore: further refine grunt tasks and their structure, update dev docs
